requirements.txt
psycopg2==2.9.10
python-dotenv
flask
flask-sqlalchemy